feat: start ws server, linter settings
